+++ src/main/java/net/djahmo/bakeddelight/registry/ModBlocks.java
@@ -1,6 +1,7 @@
 package net.djahmo.bakeddelight.registry;
 
 import net.djahmo.bakeddelight.BakedDelight;
+import net.djahmo.bakeddelight.custom.block.BackingDishTypeCollection;
 import net.djahmo.bakeddelight.custom.block.BakingCookedDishBlock;
 import net.djahmo.bakeddelight.custom.block.BakingDishBlock;
 import net.djahmo.bakeddelight.custom.block.BakingUncookedDishBlock;
@@ -20,10 +21,19 @@ import java.util.function.Supplier;
 
 public class ModBlocks {
     public static final DeferredRegister<Block> BLOCKS;
-    public static final RegistryObject<Block> BAKING_DISH, UNCOOKED_LASAGNA_DISH, COOKED_LASAGNA_DISH, UNCOOKED_GRATIN_DISH, COOKED_GRATIN_DISH;
+    public static final RegistryObject<Block> BAKING_DISH, UNCOOKED_LASAGNA_DISH, COOKED_LASAGNA_DISH, UNCOOKED_GRATIN_DISH, COOKED_GRATIN_DISH, UNCOOKED_MOUSSAKA_DISH, COOKED_MOUSSAKA_DISH;
     public static final RegistryObject<Block> EGGPLANT_CROP;
+
+    private static int nextDishId = 0;
+
     public static void register(IEventBus eventBus) {
         BLOCKS.register(eventBus);
+    }
+
+    public static RegistryObject<Block> setRecipeBlock(String name, Supplier<Block> block) {
+        nextDishId++;
+        BackingDishTypeCollection.registerDishType(name, nextDishId);
+        return setBlock(name, block);
     }
 
     public static RegistryObject<Block> setBlock(String name, Supplier<Block> block) {
@@ -41,12 +51,14 @@ public class ModBlocks {
 
         BAKING_DISH = setBlock("baking_dish", () -> new BakingDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)));
 
-        UNCOOKED_LASAGNA_DISH = setBlock("uncooked_lasagna_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)));
+        UNCOOKED_LASAGNA_DISH = setRecipeBlock("uncooked_lasagna_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)));
         COOKED_LASAGNA_DISH = setBlock("cooked_lasagna_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F), ModItems.LASAGNA_SLICE));
 
-        UNCOOKED_GRATIN_DISH = setBlock("uncooked_gratin_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)));
+        UNCOOKED_GRATIN_DISH = setRecipeBlock("uncooked_gratin_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)));
         COOKED_GRATIN_DISH = setBlock("cooked_gratin_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F), ModItems.GRATIN_SLICE));
 
+        UNCOOKED_MOUSSAKA_DISH = setRecipeBlock("uncooked_moussaka_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)));
+        COOKED_MOUSSAKA_DISH = setBlock("cooked_moussaka_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F), ModItems.MOUSSAKA_SLICE));
 
         EGGPLANT_CROP = BLOCKS.register("eggplant_crop", () -> new EggplantCropBlock(BlockBehaviour.Properties.copy(Blocks.WHEAT).sound(SoundType.CROP).noCollission().noOcclusion()));
     }