+++ changelog.md
@@ -1,6 +1,10 @@
 # Version 0.0.2
 
-
+- Add Moussaka recipe
+- Fix bug when two or more recipes have the same starting item
+- Dynamically generate recipes types, replace EnumProperty by IntegerProperty
+- Add Eggplant Crop
+- Refactor some core and reorganize the data model file
 
 # Version 0.0.1